Set backup branch to main branch
