+++ package.json
@@ -12,15 +12,16 @@
     "@popperjs/core": "^2.11.5",
     "bootstrap": "^5.1.3",
     "bootstrap-icons": "^1.8.1",
-    "chart.js": "^3.7.1",
+    "chart.js": "^2.9.4",
     "core-js": "^3.6.5",
+    "dist": "^0.1.2",
     "firebase": "^9.6.7",
     "lodash": "^4.17.21",
     "mdb-ui-kit": "^3.11.0",
     "register-service-worker": "^1.7.2",
     "three": "^0.139.2",
     "vee-validate": "^4.5.10",
-    "vue": "^3.2.33",
+    "vue": "^3.0.0",
     "vue-chartkick": "^1.1.0",
     "vue-router": "^4.0.12",
     "vuex": "^4.0.2",
