+++ src/Controllers/ReformatArrayList.js
@@ -1,7 +1,3 @@
-//  arraylist met objects te gaan naar 1 object met key values
-// from : [ {"04-04-2022": 45}  ,{"04-04-2022": 45}, {"04-04-2022": 45}  ]
-// to -->  {"04-04-2022": 45 ,"04-04-2022": 45, "04-04-2022": 45}
-
 // different days work because they have different keys
 
 export function ReformatArrayList(arraylistWithObjects) {
