+++ src/views/Feed.vue
@@ -4,11 +4,16 @@ Register.vue - base vue
   <h1>XsensDotSensor Development</h1>
 
   <p><button @click="getData()">connect</button></p>
-  <input :value="DeviceName" @change="updateDeviceName">
   <p><button @click="sync()">synchronize</button></p>
   <p><button @click="identify()">identify device</button></p>
-  <h2>Battery level: <p id="batterylevel">0</p></h2>
-  <h2>Sensor status: <p id="sensorStatus">offline</p></h2>
+  <h2>
+    Battery level:
+    <p id="batterylevel">0</p>
+  </h2>
+  <h2>
+    Sensor status:
+    <p id="sensorStatus">offline</p>
+  </h2>
   <p><button @click="startDataExport()">Export data</button></p>
   <p><button @click="streamData()">Start real time streaming</button></p>
   <p><button @click="stopDataStream()">stop real time streaming</button></p>
@@ -27,9 +32,14 @@ Register.vue - base vue
 </template>
 
 <script>
-import NavBarTop from '../components/navbars/NavBarTop.vue';
+import NavBarTop from "../components/navbars/NavBarTop.vue";
 import { XsensDotSensor } from "/src/libraries/bluetooth.js";
-import { findBluetoothDevices, startRTStream, stopRTStream, getSyncStatusSensor } from "/src/libraries/interface.js";
+import {
+  findBluetoothDevices,
+  startRTStream,
+  stopRTStream,
+  syncSensor,
+} from "/src/libraries/bluetooth.js";
 export default {
   name: "Feed",
   components: {
@@ -42,20 +52,17 @@ export default {
     };
   },
   created() {
-    window.addEventListener('beforeunload', this.handler)
+    window.addEventListener("beforeunload", this.handler);
   },
   methods: {
     getData() {
-      findBluetoothDevices()
-    },
-    updateDeviceName(e) {
-      XsensDotSensor.writeDeviceName(e.target.value.trim())
+      findBluetoothDevices();
     },
     sync() {
-      getSyncStatusSensor()
+      syncSensor();
     },
     identify() {
-      XsensDotSensor.blinkDeviceLED()
+      XsensDotSensor.blinkDeviceLED();
     },
     startDataExport() {
       XsensDotSensor.downloadDataToCSV();
