+++ vue.config.js
@@ -2,7 +2,10 @@
 
 module.exports = {
   // ...Other Vue config options here, possibly
-  publicPath: process.env.NODE_ENV === "production" ? "/project334/" : "/",
+  publicPath:
+    process.env.NODE_ENV === "production"
+      ? "/334-HOGESCHOOL-UTRECHT-INSTITUUT-VOOR-BEWEGINGSSTUDIES/"
+      : "/",
   pwa: {
     name: "project",
     themeColor: "#d8222a",
